need to create policy/cluster then relationship
